requirements.txt
fastapi~=0.117.1
uvicorn~=0.36.0
requests~=2.32.3
pillow~=11.2.1
httpx~=0.28.1